+++ src/components/Layout/index.tsx
@@ -1,8 +1,8 @@
 import type { ReactNode } from 'react';
 import Head from 'next/head';
 
-import Navbar from '@/components/Navbar/Navbar';
-import Footer from '@/components/Footer/Footer';
+import Navbar from '@/components/Navbar';
+import Footer from '@/components/Footer';
 
 interface IProps {
 	children: ReactNode;
