+++ src/pages/projects.tsx
@@ -1,7 +1,6 @@
 import Head from 'next/head';
-
-import Project from '@/components/Projects/Projects';
-import ReadyToTalk from '@/components/ReadyToTalk/ReadyToTalk';
+import Project from '@/components/Projects';
+import ReadyToTalk from '@/components/ReadyToTalk';
 import Layout from '@/components/Layout';
 export default function Home() {
 	return (
